Update solo_study.md
thursday and friday. I didn't continue on with the koans and jsexcercises first because it was confusing as to how it was written out and explained and secondly because i don't believe i had really gotten a good grip on the fundamentals. that is why you see all the stretch excercises done.